+++ examples/miniTFTWing/graphicstest/graphicstest.ino
@@ -213,7 +213,7 @@ void testdrawcircles(uint8_t radius, uint16_t color) {
 
 void testtriangles() {
   tft.fillScreen(ST77XX_BLACK);
-  int color = 0xF800;
+  uint16_t color = 0xF800;
   int t;
   int w = tft.width()/2;
   int x = tft.height()-1;
@@ -230,7 +230,7 @@ void testtriangles() {
 
 void testroundrects() {
   tft.fillScreen(ST77XX_BLACK);
-  int color = 100;
+  uint16_t color = 100;
   int x = 0;
   int y = 0;
   int w = tft.width()-2;
